requirements.txt
Django>=4.2
twilio>=9.0.0
python-dotenv>=1.0.0
